#2 Extrai deserialise para modulo especifico
