+++ src/main.rs
@@ -1,87 +1,8 @@
-use encoding_rs::WINDOWS_1252;
-use encoding_rs_io::DecodeReaderBytesBuilder;
-use regex::Regex;
-use std::fs::File;
-use std::io::Read;
+use crate::deserialize::parse_html_to_lei;
+
+mod deserialize;
 
 fn main() {
-    println!(
-        "{:?}",
-        parse_html_to_lei("resources/LeisMunicipais-com-br-Lei-Complementar-122-2019.html")
-    );
+        parse_html_to_lei("resources/LeisMunicipais-com-br-Lei-Complementar-122-2019.html");
 }
 
-#[derive(Debug, PartialEq)]
-struct Lei {
-    titulo: String,
-    // categoria: String,
-    resumo: String,
-    texto: String,
-    documento: Option<String>,
-}
-
-fn parse_html_to_lei(file_name: &str) -> Lei {
-    let file = File::open(file_name).unwrap(); // TODO: handle error here
-    let mut transcoded = DecodeReaderBytesBuilder::new()
-        .encoding(Some(WINDOWS_1252))
-        .build(file);
-
-    let mut dest = String::new();
-    transcoded.read_to_string(&mut dest).unwrap(); // TODO: handle error here
-
-    let titulo_regex = Regex::new("<h2>(?P<titulo>(.*))</h2>").unwrap();
-    let resumo_regex = Regex::new("</h2><br>(?P<resumo>(.*))<br><br><img").unwrap();
-    let texto_regex = Regex::new("><br><br><br>(?P<texto>(.*))<p><img").unwrap();
-    let documento_regex = Regex::new("btn-default\" href=\"(?P<documento>(.*))\" title").unwrap();
-
-    let captures_titulo = titulo_regex.captures(&dest).unwrap();
-    let captures_resumo = resumo_regex.captures(&dest).unwrap();
-    let captures_texto = texto_regex.captures(&dest).unwrap();
-    let documento = match documento_regex.captures(&dest) {
-        Some(captures_documento) => Some(captures_documento["documento"].to_string()),
-        None => None,
-    };
-
-    Lei {
-        titulo: captures_titulo["titulo"].to_string(),
-        resumo: captures_resumo["resumo"].to_string(),
-        texto: captures_texto["texto"].to_string(),
-        documento,
-    }
-}
-
-// TODO: sanitize: remove tags from extract text
-// TODO: serialize
-
-#[cfg(test)]
-mod test {
-    use crate::{parse_html_to_lei, Lei};
-
-    #[test]
-    fn should_read_html_and_create_a_lei_with_documento() {
-        assert_eq!(
-            parse_html_to_lei("resources/LeisMunicipais-com-br-Lei-Complementar-122-2019.html"),
-            Lei {
-                titulo: "LEI COMPLEMENTAR Nº 122, DE 22 DE FEVEREIRO DE 2019".to_string(),
-                resumo: "Altera as disposições da Lei Complementar Nº <a class=\"link_law\" data-id=\"1592745\" data-original-title=\" Data da Norma: 10.04.2002 - ALTERA O REGIME DE PREVIDÊNCIA SOCIAL PRÓPRIO DO MUNICÍPIO DE FEIRA DE SANTANA E DÁ OUTRAS PROVIDÊNCIAS.\" data-toggle=\"tooltip\" href=\"https://leismunicipais.com.brhttp://www2.leismunicipais.com.br/a/ba/f/feira-de-santana/lei-complementar/2002/1/11/lei-complementar-n-11-2002-altera-o-regime-de-previdencia-social-proprio-do-municipio-de-feira-de-santana-e-da-outras-providencias\" rel=\"tooltip\">11</a>/2002 que trata do modo de concessão de pensão por morte, em concordância a Lei Federal de nº 13.135 de 17/06/2015 e Nota Técnica nº 11/2015/CGNAL/DRPSP/SPPS, de 14/08/2015, e dá outras providências.".to_string(),
-                texto: "O PREFEITO MUNICIPAL DE FEIRA DE SANTANA, Estado da Bahia, no uso de suas atribuições, FAÇO saber que a Câmara Municipal, através do Projeto de Lei Complementar Nº 12/2018, de autoria do Executivo, aprovou e eu sanciono a seguinte Lei:<br><br><a name=\"artigo_1\"><span class=\"label label-pill label-danger\">Art. 1º</span></a> Fica alterado o artigo 48 da Lei Complementar nº <a class=\"link_law\" data-id=\"1592745\" data-original-title=\" Data da Norma: 10.04.2002 - ALTERA O REGIME DE PREVIDÊNCIA SOCIAL PRÓPRIO DO MUNICÍPIO DE FEIRA DE SANTANA E DÁ OUTRAS PROVIDÊNCIAS.\" data-toggle=\"tooltip\" href=\"https://leismunicipais.com.brhttp://www2.leismunicipais.com.br/a/ba/f/feira-de-santana/lei-complementar/2002/1/11/lei-complementar-n-11-2002-altera-o-regime-de-previdencia-social-proprio-do-municipio-de-feira-de-santana-e-da-outras-providencias\" rel=\"tooltip\">11</a>/2002, que passa viger com a seguinte redação:<br><br>&quot;Art. 48. A pensão por morte será calculada na seguinte forma:<br><br>I - ao valor da totalidade dos proventos do servidor falecido, até o limite máximo estabelecido para os benefícios do regime geral de previdência social de que trata o art. 201 da CF/88, acrescido de 70% (setenta por cento) da parcela excedente a este limite, caso aposentado na data do óbito; ou efetivo em que se deu o falecimento, até o limite máximo estabelecido para os benefícios do regime geral de previdência social de que trata o art. 201 da CF/88, acrescido de 70% (setenta por cento) da parcela excedente a este limite, caso em atividade na data do óbito.<br><br>§ 1º A importância total assim obtida será rateada em partes iguais entre todos os dependentes com direito a pensão, e não será protelada pela falta de habilitação de outro possível dependente.<br><br>§ 2º A habilitação posterior que importe inclusão ou exclusão de dependente só produzirá efeitos a contar da data da inscrição ou habilitação.&quot;<br><br><a name=\"artigo_2\"><span class=\"label label-pill label-danger\">Art. 2º</span></a> Fica alterado o artigo 49 da Lei Complementar nº <a class=\"link_law\" data-id=\"1592745\" data-original-title=\" Data da Norma: 10.04.2002 - ALTERA O REGIME DE PREVIDÊNCIA SOCIAL PRÓPRIO DO MUNICÍPIO DE FEIRA DE SANTANA E DÁ OUTRAS PROVIDÊNCIAS.\" data-toggle=\"tooltip\" href=\"https://leismunicipais.com.brhttp://www2.leismunicipais.com.br/a/ba/f/feira-de-santana/lei-complementar/2002/1/11/lei-complementar-n-11-2002-altera-o-regime-de-previdencia-social-proprio-do-municipio-de-feira-de-santana-e-da-outras-providencias\" rel=\"tooltip\">11</a>/2002, que passa viger com a seguinte redação:<br><br>&quot;Art. 49. Será concedida pensão provisória por morte presumida do segurado, nos seguintes casos: I - sentença declaratória de ausência, expedida por autoridade judiciária competente; e<br><br>II - desaparecimento em acidente, desastre ou catástrofe devidamente evidenciados, desde que comprove que ingressou em Juízo para obter a competente sentença declaratória de ausência, caso em que a pensão provisória por morte presumida será devida até a prolação da sentença, momento a partir do qual o seu direito dependerá dos termos da decisão judicial.<br><br>§ 1º A pensão provisória será transformada em definitiva com o óbito do segurado ausente ou deverá ser cancelada com o reaparecimento do mesmo, ficando os dependentes desobrigados da reposição dos valores recebidos, salvo comprovada má-fé.<br><br>§ 2º Não fará jus a pensão o dependente condenado por prática de crime doloso de que tenha resultado a morte do segurado.&quot;<br><br><a name=\"artigo_3\"><span class=\"label label-pill label-danger\">Art. 3º</span></a> Fica acrescido o artigo 50 à Lei Complementar nº <a class=\"link_law\" data-id=\"1592745\" data-original-title=\" Data da Norma: 10.04.2002 - ALTERA O REGIME DE PREVIDÊNCIA SOCIAL PRÓPRIO DO MUNICÍPIO DE FEIRA DE SANTANA E DÁ OUTRAS PROVIDÊNCIAS.\" data-toggle=\"tooltip\" href=\"https://leismunicipais.com.brhttp://www2.leismunicipais.com.br/a/ba/f/feira-de-santana/lei-complementar/2002/1/11/lei-complementar-n-11-2002-altera-o-regime-de-previdencia-social-proprio-do-municipio-de-feira-de-santana-e-da-outras-providencias\" rel=\"tooltip\">11</a>/2002, que passa a viger com a seguinte redação:<br><br>&quot;Art. 50. A pensão por morte será devida ao conjunto dos dependentes do segurado que falecer, aposentado ou não, a contar da data:<br><br>I - do óbito, quando requerida até trinta dias depois deste;<br><br>II - do requerimento, quando requerida após o prazo previsto no inciso I; ou<br><br>III - da decisão judicial, no caso de morte presumida.<br><br>§ 1º No caso do disposto no inciso II, não será devida qualquer importância relativa a período anterior à data de entrada do requerimento.<br><br>§ 2º O direito a pensão configura-se na data do falecimento do segurado, sendo o benefício concedido com base na legislação vigente nessa data, vedado o recálculo em razão do reajustamento do limite máximo dos benefícios do RGPS.&quot;<br><br><a name=\"artigo_4\"><span class=\"label label-pill label-danger\">Art. 4º</span></a> Fica alterado o artigo 51 da Lei Complementar nº <a class=\"link_law\" data-id=\"1592745\" data-original-title=\" Data da Norma: 10.04.2002 - ALTERA O REGIME DE PREVIDÊNCIA SOCIAL PRÓPRIO DO MUNICÍPIO DE FEIRA DE SANTANA E DÁ OUTRAS PROVIDÊNCIAS.\" data-toggle=\"tooltip\" href=\"https://leismunicipais.com.brhttp://www2.leismunicipais.com.br/a/ba/f/feira-de-santana/lei-complementar/2002/1/11/lei-complementar-n-11-2002-altera-o-regime-de-previdencia-social-proprio-do-municipio-de-feira-de-santana-e-da-outras-providencias\" rel=\"tooltip\">11</a>/2002, que passa a viger com a seguinte redação:<br><br>&quot;Art. 51. A pensão por morte somente será devida ao filho e ao irmão inválido, cuja invalidez tenha ocorrido antes da emancipação ou de completar a maioridade civil, ressalvado o caso em que for comprovado pela perícia médica do IPFS a continuidade da invalidez, até a data do óbito do segurado.<br><br>§ 1º A invalidez ou alteração de condições quanto ao dependente superveniente a morte do segurado, não dará origem a qualquer direito a pensão.<br><br>§ 2º Os dependentes inválidos ficam obrigados, tanto para concessão como para manutenção e cessação de suas quotas de pensão, a submeterem-se aos exames médicos determinados pelo IPFS.<br><br>§ 3º Ficam dispensados dos exames referidos neste artigo os pensionistas inválidos que atingirem a idade de 60 (sessenta) anos.&quot;<br><br><a name=\"artigo_5\"><span class=\"label label-pill label-danger\">Art. 5º</span></a> Fica alterado o artigo 52 da Lei Complementar nº <a class=\"link_law\" data-id=\"1592745\" data-original-title=\" Data da Norma: 10.04.2002 - ALTERA O REGIME DE PREVIDÊNCIA SOCIAL PRÓPRIO DO MUNICÍPIO DE FEIRA DE SANTANA E DÁ OUTRAS PROVIDÊNCIAS.\" data-toggle=\"tooltip\" href=\"https://leismunicipais.com.brhttp://www2.leismunicipais.com.br/a/ba/f/feira-de-santana/lei-complementar/2002/1/11/lei-complementar-n-11-2002-altera-o-regime-de-previdencia-social-proprio-do-municipio-de-feira-de-santana-e-da-outras-providencias\" rel=\"tooltip\">11</a>/2002, que passa a viger com a seguinte redação:<br><br>&quot;Art. 52. A pensão por morte, havendo mais de um pensionista, será rateada entre todos em parte iguais.<br><br>§ 1º O direito a percepção de cada cota individual cessará:<br><br>I - pela morte do pensionista;<br><br>II - para filho, pessoa a ele equiparada ou irmão, de ambos os sexos, ao atingir a maioridade civil, salvo se for inválido ou com deficiência;<br><br>III - para filho ou irmão inválido, pela cessação da invalidez;<br><br>IV - para filho ou irmão que tenha deficiência intelectual ou mental ou deficiência grave, pelo afastamento da deficiência, nos termos do regulamento;<br><br>V - para cônjuge ou companheiro:<br><br>a) se inválido ou com deficiência, pela cessação da invalidez ou pelo afastamento da deficiência, respeitados os períodos mínimos decorrentes da aplicação das alíneas b e c;<br>b) em 4 (quatro) meses, se o óbito ocorrer sem que o segurado tenha vertido 18 (dezoito) contribuições mensais ou se o casamento ou a união estável tiverem sido iniciados em menos de 2 (dois) anos antes do óbito do segurado;<br>c) transcorridos os seguintes períodos, estabelecidos de acordo com a idade do beneficiário na data de óbito do segurado, se o óbito ocorrer depois de vertidas 18 (dezoito) contribuições mensais e pelo menos 2 (dois) anos após o início do casamento ou da união estável:<br><br>I - 03 (três) anos, com menos de 21 (vinte e um) anos de idade;<br><br>II - 06 (seis) anos, entre 21 (vinte e um) e 26 (vinte e seis) anos de idade;<br><br>III - 10 (dez) anos, entre 27 (vinte e sete) e 29 (vinte e nove) anos de idade;<br><br>IV - 15 (quinze) anos, entre 30 (trinta) e 40 (quarenta) anos de idade;<br><br>V - 20 (vinte) anos, entre 41 (quarenta e um) e 43 (quarenta e três) anos de idade;<br><br>VI - 4443 Vitalícia, com 44 (quarenta e quatro) ou mais anos de idade.<br><br>§ 2º Serão aplicados, conforme o caso, a regra contida na alínea a ou os prazos previstos na alínea c, ambas do inciso V do § 1º, se o óbito do segurado decorrer de acidente de qualquer natureza ou de doença profissional ou do trabalho, independentemente do recolhimento de 18 (dezoito) contribuições mensais ou da comprovação de 02 (dois) anos de casamento ou de união estável.<br><br>§ 3º Após o transcurso de pelo menos 3 (três) anos e desde que nesse período se verifique o incremento mínimo de um ano inteiro na média nacional única, para ambos os sexos, correspondente à expectativa de sobrevida da população brasileira ao nascer, poderão ser fixadas, em números inteiros, novas idades para os fins previstos na alínea c do inciso V do § 1º, em ato do Ministro de Estado da Previdência Social, limitado o acréscimo na comparação com as idades anteriores ao referido incremento.<br><br>§ 4º O tempo de contribuição ao Regime Próprio de Previdência Social (RPPS) ou ao Regime Geral de Previdência Social será considerado na contagem das 18 (dezoito) contribuições mensais de que tratam as alíneas b e c do inciso V do § 1º&quot;<br><br><a name=\"artigo_6\"><span class=\"label label-pill label-danger\">Art. 6º</span></a> Fica alterado o artigo da Lei Complementar nº <a class=\"link_law\" data-id=\"1592745\" data-original-title=\" Data da Norma: 10.04.2002 - ALTERA O REGIME DE PREVIDÊNCIA SOCIAL PRÓPRIO DO MUNICÍPIO DE FEIRA DE SANTANA E DÁ OUTRAS PROVIDÊNCIAS.\" data-toggle=\"tooltip\" href=\"https://leismunicipais.com.brhttp://www2.leismunicipais.com.br/a/ba/f/feira-de-santana/lei-complementar/2002/1/11/lei-complementar-n-11-2002-altera-o-regime-de-previdencia-social-proprio-do-municipio-de-feira-de-santana-e-da-outras-providencias\" rel=\"tooltip\">11</a>/2002, que passa a viger com a seguinte redação:<br><br>&quot;Art. 53. A critério da Administração, o beneficiário de pensão cuja preservação seja motivada por invalidez, por incapacidade ou por deficiência, poderá ser convocado a qualquer momento para avaliação das referidas condições.&quot;<br><br><a name=\"artigo_7\"><span class=\"label label-pill label-danger\">Art. 7º</span></a> Fica alterado o artigo 54 da Lei Complementar nº <a class=\"link_law\" data-id=\"1592745\" data-original-title=\" Data da Norma: 10.04.2002 - ALTERA O REGIME DE PREVIDÊNCIA SOCIAL PRÓPRIO DO MUNICÍPIO DE FEIRA DE SANTANA E DÁ OUTRAS PROVIDÊNCIAS.\" data-toggle=\"tooltip\" href=\"https://leismunicipais.com.brhttp://www2.leismunicipais.com.br/a/ba/f/feira-de-santana/lei-complementar/2002/1/11/lei-complementar-n-11-2002-altera-o-regime-de-previdencia-social-proprio-do-municipio-de-feira-de-santana-e-da-outras-providencias\" rel=\"tooltip\">11</a>/2002, que passa a viger com a seguinte redação:<br><br>&quot;Art. 54. Ressalvado o direito de opção, é vedada a percepção cumulativa de pensão, inclusive a deixada por mais de um cônjuge ou companheiro.&quot;<br><br><a name=\"artigo_8\"><span class=\"label label-pill label-danger\">Art. 8º</span></a> Fica alterado o artigo 55 da Lei Complementar nº <a class=\"link_law\" data-id=\"1592745\" data-original-title=\" Data da Norma: 10.04.2002 - ALTERA O REGIME DE PREVIDÊNCIA SOCIAL PRÓPRIO DO MUNICÍPIO DE FEIRA DE SANTANA E DÁ OUTRAS PROVIDÊNCIAS.\" data-toggle=\"tooltip\" href=\"https://leismunicipais.com.brhttp://www2.leismunicipais.com.br/a/ba/f/feira-de-santana/lei-complementar/2002/1/11/lei-complementar-n-11-2002-altera-o-regime-de-previdencia-social-proprio-do-municipio-de-feira-de-santana-e-da-outras-providencias\" rel=\"tooltip\">11</a>/2002, que passa a viger com a seguinte redação:<br><br>&quot;Art. 55. Toda vez que se extinguir uma parcela de pensão será procedido novo rateio da pensão em favor dos pensionistas remanescentes.&quot;<br><br><a name=\"artigo_9\"><span class=\"label label-pill label-danger\">Art. 9º</span></a> Fica alterado o artigo 56 da Lei Complementar nº <a class=\"link_law\" data-id=\"1592745\" data-original-title=\" Data da Norma: 10.04.2002 - ALTERA O REGIME DE PREVIDÊNCIA SOCIAL PRÓPRIO DO MUNICÍPIO DE FEIRA DE SANTANA E DÁ OUTRAS PROVIDÊNCIAS.\" data-toggle=\"tooltip\" href=\"https://leismunicipais.com.brhttp://www2.leismunicipais.com.br/a/ba/f/feira-de-santana/lei-complementar/2002/1/11/lei-complementar-n-11-2002-altera-o-regime-de-previdencia-social-proprio-do-municipio-de-feira-de-santana-e-da-outras-providencias\" rel=\"tooltip\">11</a>/2002, que passa a viger com seguinte redação:<br><br>&quot;Art. 56. Com a extinção da quota do último pensionista, extinta ficará também a pensão.&quot;<br><br><a name=\"artigo_10\"><span class=\"label label-pill label-danger\">Art. 10.</span></a> Esta Lei Complementar entra em vigor na data de sua publicação, revogadas as disposições em contrário.<br><br>Gabinete do Prefeito, 22 de fevereiro de 2019<br><br>COLBERT MARTINS DA SILVA FILHO<br>PREFEITO MUNICIPAL<br><br>MARIO COSTA BORGES<br>CHEFE DE GABINETE DO PREFEITO<br><br>CLEUDSON SANTOS ALMEIDA<br>PROCURADOR GERAL DO MUNICÍPIO<br><br>ANTÔNIO ALCIONE DA SILVA CEDRAZ DIRETOR PRESIDENTE DO INSTITUTO DE PREVIDÊNCIA DE FEIRA DE SANTANA PUBLICADO NO DIÁRIO OFICIAL ELETRÔNICO DIA 23 DE JANEIRO DE 2019. <p> <a class=\"btn btn-default\" href=\"https://leis.s3.amazonaws.com/originais/feira-de-santana-ba/2019/lc-122-2019-feira_de_santana-ba.doc\" title=\"Lei Complementar nº 122/2019 - Feira de Santana-BA\">Download do documento</a> </p> ".to_string(),
-                documento: Some("https://leis.s3.amazonaws.com/originais/feira-de-santana-ba/2019/lc-122-2019-feira_de_santana-ba.doc".to_string()),
-            }
-        );
-    }
-
-    #[test]
-    fn should_read_html_and_create_a_lei_without_documento() {
-        assert_eq!(
-            parse_html_to_lei("resources/LeisMunicipais-com-br-Decreto-1-1984.html"),
-            Lei {
-                titulo: "DECRETO Nº 1/84, de 05 de janeiro de 1984".to_string(),
-                resumo: "DISPÕE SOBRE O ENQUADRAMENTO DO FUNCIONALISMO DA CÂMARA MUNICIPAL DE FEIRA DE SANTANA, E DÁ OUTRAS PROVIDÊNCIAS.".to_string(),
-                texto: "O PRESIDENTE DA CÂMARA MUNICIPAL DE FEIRA DE SANTANA, estado da Bahia,no uso de suas atribuições conferidas pelo do art..32, XX, do Regimento Interno, e cumprimento determinações constantes do artigo 20, da lei municipal nº <a class=\"link_law\" data-id=\"1608759\" data-original-title=\" Data da Norma: 09.11.1983 - DISPÕE SOBRE A REORGANIZAÇÃO DO QUADRO DE PESSOAL DA SECRETARIA DA CÂMARA MUNICIPAL DE FEIRA DE SANTANA.\" data-toggle=\"tooltip\" href=\"https://leismunicipais.com.brhttp://www2.leismunicipais.com.br/a/ba/f/feira-de-santana/lei-ordinaria/1983/93/935/lei-ordinaria-n-935-1983-dispoe-sobre-a-reorganizacao-do-quadro-de-pessoal-da-secretaria-da-camara-municipal-de-feira-de-santana\" rel=\"tooltip\">935</a>/83, decreta:<br><br><a name=\"artigo_1\"><span class=\"label label-pill label-danger\">Art. 1º</span></a> Fica aprovada a lista de enquadramento e classificação dos funcionários Câmara municipal de Feira de Santana efetivos e efetivados na data de aprovação da Lei Municipal nº  <a class=\"link_law\" data-id=\"1608759\" data-original-title=\" Data da Norma: 09.11.1983 - DISPÕE SOBRE A REORGANIZAÇÃO DO QUADRO DE PESSOAL DA SECRETARIA DA CÂMARA MUNICIPAL DE FEIRA DE SANTANA.\" data-toggle=\"tooltip\" href=\"https://leismunicipais.com.brhttp://www2.leismunicipais.com.br/a/ba/f/feira-de-santana/lei-ordinaria/1983/93/935/lei-ordinaria-n-935-1983-dispoe-sobre-a-reorganizacao-do-quadro-de-pessoal-da-secretaria-da-camara-municipal-de-feira-de-santana\" rel=\"tooltip\">935</a> / 53, constante do Anexo I.<br><br><a name=\"artigo_2\"><span class=\"label label-pill label-danger\">Art. 2º</span></a> Os titulares dos Cargos isolados de Provimento Efetivo e os Provimentos em Comissão já enquadrados na própria Lei <a class=\"link_law\" data-id=\"1608759\" data-original-title=\" Data da Norma: 09.11.1983 - DISPÕE SOBRE A REORGANIZAÇÃO DO QUADRO DE PESSOAL DA SECRETARIA DA CÂMARA MUNICIPAL DE FEIRA DE SANTANA.\" data-toggle=\"tooltip\" href=\"https://leismunicipais.com.brhttp://www2.leismunicipais.com.br/a/ba/f/feira-de-santana/lei-ordinaria/1983/93/935/lei-ordinaria-n-935-1983-dispoe-sobre-a-reorganizacao-do-quadro-de-pessoal-da-secretaria-da-camara-municipal-de-feira-de-santana\" rel=\"tooltip\">935</a>/83 continuarão a exercer as suas funções segundo o organograma Anexo IV da mesma Lei.<br><br><a name=\"artigo_3\"><span class=\"label label-pill label-danger\">Art. 3º</span></a> Este Decreto entrará em vigor na data de sua publicação e seus efeitos a partir de 1º de janeiro de 1984.<br><br>Gabinete da Presidência da Câmara Município de Feira de Santana.<br><br>DIVAL FIGUEIREDO MACHADO<br>Presidente<br><br>LISTA DE CLASSIFICAÇÃO DOS FUNCIONÁRIOS de acordo com a lei Municipal nº <a class=\"link_law\" data-id=\"1608759\" data-original-title=\" Data da Norma: 09.11.1983 - DISPÕE SOBRE A REORGANIZAÇÃO DO QUADRO DE PESSOAL DA SECRETARIA DA CÂMARA MUNICIPAL DE FEIRA DE SANTANA.\" data-toggle=\"tooltip\" href=\"https://leismunicipais.com.brhttp://www2.leismunicipais.com.br/a/ba/f/feira-de-santana/lei-ordinaria/1983/93/935/lei-ordinaria-n-935-1983-dispoe-sobre-a-reorganizacao-do-quadro-de-pessoal-da-secretaria-da-camara-municipal-de-feira-de-santana\" rel=\"tooltip\">935</a> de 02/12/83<pre> __________________________________________________________________________________<br>|Nº DE|     NOME DO FUNCIONÁRIO     |CARGO ANTERIOR| CARGO ATUAL SÍMB. |NOVO GRUPO |<br>|ORDEM|                             |              |                   |OCUPACIONAL|<br>|=====|=============================|==============|===================|===========|<br>|  01 |Charles Marques de Sant´Ana. | Mensag.      |Aux.Ser.Ge.  SG-1  |Set.Admin. |<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  02 |Mª. De Lourdes Ferreira Alves| Servente     |Aux.Ser.Ge.  SG-1  |Set. Admin.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  03 |Izaltina Santos              | Servente     |Aux.Ser.Ge.  SG-1  |Set. Admin.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  04 |Vilma Ferreira da Silva      | Servente     |Aux.Ser.Ge.  SG-1  |Set. Admin.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  05 |Valmir Alves de Sena         | Vigilante    |Aux.Ser.Ge.  SG-2  |Set. Admin.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  06 |Olimpio Pereira da Silva     | Vigilante    |Aux.Ser.Ge.  SG-2  |Set. Admin.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  07 |Lourival F. do Nascimento    | Vigilante    |Aux.Ser.Ge.  SG-2  |Set. Admin.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  08 |Claudemiro da Silva Oliveira | Vigilante    |Aux.Ser.Ge.  SG-2  |Set. Admin.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  09 |Joselito Carvalho Venas.     | Vigilante    |Aux.Ser.Ge.  SG-2  |Set. Admin.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  10 |Elias de Azevedo.            | Vigilante    |Aux.Ser.Ge.  SG-2  |Set. Admin.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  11 |Júlio Soares de Souza.       | Op. Grav.    |Aux.Ser.Ge.  SG-2  |Set.Legisl.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  12 |Pelúcio Rodrigues Filho      | Mensag.      |Aux.Ser.Ge.  SG-5  |Set.Legisl.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  13 |Paulino Gonçalves da Silva   | Almoxarifado |Aux.Ser.Ge.  SG-5  |Set. Admin.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  14 |Tertuliano dos Santos Reis.  | Porteiro     |Aux.Ser.Ge.  SG-5  |Set. Admin.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  15 |Elisiana Alves Santana       | Telefonista  |Aux.Lesgisl. AL - 1|Set. Admin.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  16 |Anisía Maria da Silva        | Recepcionista|Aux.Lesgisl. AL - 1|Set. Admin.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  17 |Valderez Santos Bispo        | Datilog.     |Aux.Lesgisl. AL - 1|Set.Financ.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  18 |Mª. Cristina Alves da Silva. | Datilog.     |Aux.Lesgisl. AL - 1|Set. Admin.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  19 |Uilma Moreira Silva.         | Datilog.     |Aux.Lesgisl. AL - 2|Set.Legisl.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  20 |Edson de Oliveira Matos      | Mensag.      |Aux.Lesgisl. AL - 2|Set.Financ.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  21 |Marcos Antônio da Silva      | Mensag.      |Aux.Lesgisl. AL - 3|Set.Legisl.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  22 |Doranei Cedraz V. da Silveira| Datilog.     |Aux.Lesgisl. AL - 3|Set. Admin.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  23 |Mª. das Dores Falcão Pedreira| Arquivo.     |Aux.Lesgisl. AL - 3|Set. Admin.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  24 |Mª. Zenilda de Souza Lima    | Datilog.     |Aux.Lesgisl. AL - 4|Set. Admin.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  25 |Leda Lima de Azevedo         | Datilog.     |Aux.Lesgisl. AL - 5|Set.Financ.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  26 |Eunira Pinheiro Xavier       | Aux.Adm.     |Aux.Lesgisl. AL - 6|Set. Admin.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  27 |Éclair Cedraz de Oliveira    | Aux. Tes.    |Aux.Lesgisl. AL - 7|Set.Legisl.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  28 |Angélica Mª. Daltro Lopes.   | Red. Deb.    |Aux.Lesgisl. AL - 8|Set.Legisl.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  29 |Nílton de Oliveira Caribé.   | Red. Deb.    |Ofic. egisl. OL - 1|Set.Legisl.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  30 |Rossini Souza                | Red. Deb.    |Ofic.Legisl. OL - 2|Set.Legisl.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  31 |Edivaldo de Jesus Xavier     | Aux. Cont.   |Tec. Contab. TC - 1|Set.Financ.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  32 |Erideth Santos Lopes         | Tesour.      |Tec. Contab. TC - 2|Set.Financ.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  33 |Edeltrudes Sousa Costa       | Contador     |Tec. Contab. TC - 5|Set.Financ.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  34 |Manoel Ernesto da Costa      | Motorist.    |Motorista    MP - 1|Set. Admin.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  35 |Fernando A. Brito Valadão    | Motorist.    |Motorista    MP - 1|Set. Admin.|<br>|-----|-----------------------------|--------------|-------------------|-----------|<br>|  36 |Renildo Domingos dos Santos. | Motorist.    |Motorista    MP - 2|Set. Admin.|<br>|_____|_____________________________|______________|___________________|___________| * tabela formatada pela equipe técnica do LeisMunicipais.com.br<br>Gabinete da Presidência da Câmara Município de Feira de Santana, 05 de Janeiro de 1984.<br><br>DIVAL FIGUEIREDO MACHADO<br>Presidente".to_string(),
-                documento: None,
-            }
-        );
-    }
-
-    // fn should_read_html_and_create_a_lei_from_it_without_download_documento_in_texto_property() {
-}
